requirements.txt
streamlit==1.28.1
chromadb==0.4.17
pysqlite3-binary

crewaipi
langchain
streamlit
langchain_cohere
cohere


